requirements.txt
aiohttp
aiodns
beautifulsoup4
cchardet
elasticsearch
pysocks
pandas>=0.23.0
aiohttp_socks
schedule
geopy
fake-useragent
googletransx
notebook
nest_asyncio
hydra-core
psycopg2
newspaper3k
feedparser
peewee
elasticsearch-dsl>=7.0.0,<8.0.0
pytest
requests
requests-cache